Courses endpoint in API created

+++ src/main/java/org/rest/model/Course.java
@@ -4,17 +4,17 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class Course {
-    private long id;
+    private int id;
     private String name;
     private String lecturer;
 
-    private static long idNumber = 0;
+    private static int idNumber = 0;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
